requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.3
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ledger==0.5.6.2
django-timezone-field==6.1.0
django-treebeard==4.7.1
Faker==24.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.5
Markdown==3.5.2
ofxtools==0.9.5
pillow==10.2.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
